Moved the const check into the symbol resolution
We require symbols to be constant as if their value can change then we
don't know about what they are pointing at. An exception that will need
to be added when supported for structs is the struct symbol can be
non-const providing the component being accessed is - this is covered by
test 28.