Adding tests for different way FPs can be called
FPs may not always be assigned into a const FP and instead may be called
in place (e.g. directly from the array). This modifies the exprt
structure that must be unpacked. These tests demonstrate the different
ways that can be done. Most are KNOWNBUG as they relate to existing
KNOWNBUGs called in new ways. The ones that are not currently don't
work.